refactor(mcp): extract worker modules
