+++ src/mcp/tools/index.ts
@@ -1,4 +1,33 @@
 export {
+    executeAnalyzeDependencies,
+    executeParseImports,
+    executeVerifyDependencyUsage
+} from "./analysis";
+export {
+    executeGetSymbolCallers,
+    executeTraceFunctionExecution
+} from "./execution";
+export {
+    executeCrawlDependencyGraph,
+    executeExpandNode,
+    executeFindReferencingFiles
+} from "./graph";
+export {
+    executeAnalyzeBreakingChanges,
+    executeGetImpactAnalysis
+} from "./impact";
+export {
+    executeAnalyzeFileLogic
+} from "./logic";
+export {
+    executeResolveModulePath
+} from "./resolve";
+export {
+    executeFindUnusedSymbols,
+    executeGetSymbolDependents,
+    executeGetSymbolGraph
+} from "./symbol";
+export {
     executeGetIndexStatus,
     executeInvalidateFiles,
     executeRebuildIndex
